Rename ImageOutput to ImageOutputBeamer
