+++ src/components/image/ImageOutputBeamer.vue
