+++ src/components/presentation-types.js
@@ -57,7 +57,7 @@ export default [
       control: require('./image/ImageControl.vue').default
     },
     outputs: {
-      beamer: require('./image/ImageOutput.vue').default,
+      beamer: require('./image/ImageOutputBeamer.vue').default,
       livestream: require('./image/ImageOutputLivestream.vue').default
     }
   },
